Updated launchers and scripts to work with multiple robots
The patch solves issue #7 related to simulating multiple robots in the same environment. Each new robot now has to be started inside a specific namespace, with all its nodes, topics and services.  For this reason, both youbot_control and youbot_teleop can be started using a namespace argument, specifying which robot's namespace to work in.

The patch does not modify normal use in case the user has no interest in spawning multiple robots. In this case, normal command flow can be followed.